+++ front/app/containers/Admin/reporting/components/ReportBuilder/Editor/index.tsx
@@ -7,7 +7,6 @@ import Container from 'components/admin/ContentBuilder/Widgets/Container';
 
 import CommunityMonitorTemplate from '../Templates/CommunityMonitorTemplate';
 import PhaseTemplate from '../Templates/PhaseTemplate';
-import Insights from '../Templates/PhaseTemplate/Insights';
 import PlatformTemplate from '../Templates/PlatformTemplate';
 import ProjectTemplate from '../Templates/ProjectTemplate';
 import { WIDGETS } from '../Widgets';
@@ -27,7 +26,7 @@ const resolver = {
   ProjectTemplate,
   CommunityMonitorTemplate,
   PhaseTemplate,
-  Insights,
+  Insights: () => <div />,
   PlatformTemplate,
 };
